Small code tidying changes from full PR branch review
